Deduplicate utility functions into utildos.c/utldos2.c
