+++ src/dosfunc.c
@@ -2,7 +2,7 @@
 #include "inttype.h"
 #include "output.h"
 #include "memory.h"
-#include "util.h"
+#include "f15util.h"
 #include "offsets.h"
 
 #include <STDIO.H>
